# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateEventRequest:
    properties:
      author-id:
        type: string
      body:
        maxLength: 150
        type: string
      game:
        type: string
      max:
        type: integer
      minute:
        type: integer
    required:
    - author-id
    - game
    - max
    - minute
    type: object
...
  dto.EventResponse:
    properties:
      author-id:
        type: string
      id:
        type: string
      time:
        type: string
    type: object
  dto.JoinToEventRequest:
    properties:
      event-id:
        type: string
      user-id:
        type: string
    required:
    - event-id
    - user-id
    type: object
...
  dto.UnjoinFromEventRequest:
    properties:
      event-id:
        type: string
      user-id:
        type: string
    required:
    - event-id
    - user-id
    type: object
...
  entities.Event:
    properties:
      author_id:
        type: string
      body:
        type: string
      event_id:
        type: string
      game:
        type: string
      max:
        type: integer
      minute:
        type: string
      notifiedPre:
        type: boolean
    type: object
...
host: localhost:1111
info:
  contact: {}
  description: Ненавижу эту хуйню бля
  title: На жри уебок доки свои не подавись (недоделанная кстати хихиххихихих))))))
  version: "1.0"
paths:
...
  /events:
    get:
      description: Returns a list of events with pagination
      parameters:
      - description: amount
        in: query
        name: amount
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Event'
            type: array
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: Getting a list of events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Creates a new event
      parameters:
      - description: Event data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create an event
      tags:
      - events
  /events/{id}:
    delete:
      description: Deletes an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deleting an event
      tags:
      - events
    get:
      description: Returns an event by its id
      parameters:
      - description: event Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Event'
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      summary: Getting event by ID
      tags:
      - events
  /events/join:
    post:
      consumes:
      - application/json
      description: Adds a user to the event participants
      parameters:
      - description: Data for join to event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinToEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Joining the event
      tags:
      - events
  /events/unjoin:
    post:
      consumes:
      - application/json
      description: Removes a user from the event participants
      parameters:
      - description: Data for unjoin from event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UnjoinFromEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\":\"string\"}'
          schema:
            type: object
        "400":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "408":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
        "500":
          description: '{\"error\":\"string\"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Exit event
      tags:
      - events
...
swagger: "2.0"